Revert "Feature/loadmore functionality- when API doesnot have a Pagination support"
