+++ playground/src/App.css
@@ -1,14 +1,12 @@
 .App {
-  width:90%
+
 }
 
 ul {
+  list-style-type: none;
   padding: 0;
   width: 600px;
   margin: 0 auto;
-  text-align: left;
-}
-li {
-  padding: 0.5rem;
-  margin: 0.5rem 0;
+  display: flex;
+  justify-content: space-around
 }
