+++ playground/src/App.tsx
@@ -17,9 +17,6 @@ function App() {
         <li>
           <Link to="/query-test">Query Test</Link>
         </li>
-        <li>
-          <Link to="/load-more">Load More Sample</Link>
-        </li>
       </ul>
     </div>
   );
