--- playground/src/load-more/index.css
@@ -1,31 +1,0 @@
-.post {
-  border: 1px solid black;
-  border-radius: 10px;
-  background-color: rgb(211, 218, 195);
-  padding: 0.25rem 0.5rem;
-  width: 60%;
-  text-align: left;
-  margin: 0.25rem 0;
-}
-
-.postList {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-h4{
-    font-weight: bold;
-    font-size: 1.1rem;
-    padding: 0 0;
-    margin: 0.15rem 0;
-}
-
-.loadMore button{
-  background: none!important;
-  border: none;
-  padding: 0!important;
-  color: #069;
-  text-decoration: underline;
-  cursor: pointer;
-}
